+++ src/info/Defs.lua
@@ -1,4 +1,4 @@
----@alias BrakeData {maxDeceleration:number, currentDeceleration:number, pid:number}
+---@alias BrakeData {maxDeceleration:number, currentDeceleration:number, pid:number, setAutoBrakeAngle:number, autoBrakeAngle:number}
 
 ---@alias FlightData {targetSpeed:number, targetSpeedReason:string, finalSpeed:number, finalSpeedDistance:number, distanceToAtmo:number, atmoDistance:number, brakeMaxSpeed:number, waypointDist:number, speedDiff:number, pid:number, fsmState:string, acceleration:number, controlAcc:number, absSpeed:number}
 ---@alias AdjustmentData {long:number, lat:number, ver:number}